// src/app/models/customer.model.ts
export interface CustomerTrn {
    id?: number;
    trbactno?: number;
    trndate: string;
    customerName: string;
    toolingdrawingpartno: string;
    partdrawingname: string;
    partdrawingno: string;
    descriptionoftooling: string;
    cmworkorderno: string;
    toolingassetno: string;
    createdby: string;
    status?: string;
}